% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{data.link}
\alias{data.link}
\title{Link to the data in the MCBE package}
\usage{
data.link(x)
}
\arguments{
\item{x}{name of the file}
}
\value{
The output is the path to the file in the local system
}
\description{
Link to the data in the MCBE package
}
\examples{
data.link ("callMoney.xlsx")
}
