% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Hello my world function}
\usage{
hello(x)
}
\arguments{
\item{x}{is the name of the person to say hi}
}
\value{
The output from \code{\link{print}}
}
\description{
Hello my world function
}
\examples{
hello("Jhon")
\dontrun{
hello("Hasu")
}
}
